+++ src/fit_types.ts
@@ -474,10 +474,10 @@ export interface ParsedSession {
   max_pos_vertical_speed?: number
   max_neg_vertical_speed?: number
   min_heart_rate?: number
-  time_in_hr_zone?: number[][]
-  time_in_speed_zone?: number[][]
-  time_in_cadence_zone?: number[][]
-  time_in_power_zone?: number[][]
+  time_in_hr_zone?: number[]
+  time_in_speed_zone?: number[]
+  time_in_cadence_zone?: number[]
+  time_in_power_zone?: number[]
   avg_lap_time?: number
   best_lap_index?: number
   min_altitude?: number
@@ -630,10 +630,10 @@ export interface ParsedLap {
   avg_neg_vertical_speed?: number
   max_pos_vertical_speed?: number
   max_neg_vertical_speed?: number
-  time_in_hr_zone?: number[][]
-  time_in_speed_zone?: number[][]
-  time_in_cadence_zone?: number[][]
-  time_in_power_zone?: number[][]
+  time_in_hr_zone?: number[]
+  time_in_speed_zone?: number[]
+  time_in_cadence_zone?: number[]
+  time_in_power_zone?: number[]
   repetition_num?: number
   min_altitude?: number
   min_heart_rate?: number
@@ -880,7 +880,7 @@ export interface ParsedSchedule {
   scheduled_time?: LocalDateTime
 }
 export interface ParsedWeightScale {
-  weight?: Weight
+  weight?: number
   percent_fat?: number
   percent_hydration?: number
   visceral_fat_mass?: number
@@ -1004,7 +1004,7 @@ export interface ParsedMonitoring {
   distance16?: number
   cycles16?: number
   active_time16?: number
-  local_timestamp?: number
+  local_timestamp?: LocalDateTime
   temperature?: number
   temperature_min?: number
   temperature_max?: number
@@ -1022,7 +1022,7 @@ export interface ParsedMonitoring {
   moderate_activity_minutes?: number
   vigorous_activity_minutes?: number
   vigorous_activity_inutes?: number
-  timestamp: number
+  timestamp: string
 }
 export interface ParsedTrainingFile {
   type?: File
@@ -1033,7 +1033,7 @@ export interface ParsedTrainingFile {
   timestamp: string
 }
 export interface ParsedHrv {
-  time?: number[][]
+  time?: number[]
 }
 export interface ParsedUserMetrics {
   vo2_max?: number
@@ -1099,7 +1099,7 @@ export interface ParsedLength {
   message_index?: MessageIndex
 }
 export interface ParsedMonitoringInfo {
-  local_timestamp?: number
+  local_timestamp?: LocalDateTime
   activity_type?: ActivityType[]
   cycles_to_distance?: number[]
   cycles_to_calories?: number[]
@@ -1493,7 +1493,7 @@ export interface ParsedExdDataConceptConfiguration {
 export interface ParsedFieldDescription {
   developer_data_index?: number
   field_definition_number?: number
-  fit_base_type_id?: number
+  fit_base_type_id?: FitBaseType
   field_name?: string
   array?: number
   components?: string
@@ -1508,7 +1508,7 @@ export interface ParsedFieldDescription {
 }
 export interface ParsedDeveloperDataId {
   developer_id?: number[]
-  application_id?: number[][]
+  application_id?: number[]
   manufacturer_id?: Manufacturer
   developer_data_index?: number
   application_version?: number
@@ -1544,16 +1544,16 @@ export interface ParsedMonitoringHrData {
   timestamp: string
 }
 export interface ParsedTimeInZone {
-  reference_mesg?: number
-  reference_index?: number
-  time_in_hr_zone?: number[][]
-  time_in_speed_zone?: number[][]
-  time_in_cadence_zone?: number[][]
-  time_in_power_zone?: number[][]
-  hr_zone_high_boundary?: number[][]
-  speed_zone_high_boundary?: number[][]
-  cadence_zone_high_boundary?: number[][]
-  power_zone_high_boundary?: number[][]
+  reference_mesg?: MesgNum
+  reference_index?: MessageIndex
+  time_in_hr_zone?: number[]
+  time_in_speed_zone?: number[]
+  time_in_cadence_zone?: number[]
+  time_in_power_zone?: number[]
+  hr_zone_high_boundary?: number[]
+  speed_zone_high_boundary?: number[]
+  cadence_zone_high_boundary?: number[]
+  power_zone_high_boundary?: number[]
   hr_calc_type?: HrZoneCalc
   max_heart_rate_deprecated?: number
   max_heart_rate?: number
@@ -1569,8 +1569,8 @@ export interface ParsedSet {
   weight?: number
   set_type?: SetType
   start_time: string
-  category?: ExerciseCategory[][]
-  category_subtype?: number[][]
+  category?: ExerciseCategory[]
+  category_subtype?: number[]
   weight_display_unit?: FitBaseUnit
   message_index?: MessageIndex
   wkt_step_index?: MessageIndex
@@ -1822,12 +1822,12 @@ export interface ParsedClimbPro {
   timestamp: string
 }
 export interface ParsedTankUpdate {
-  sensor?: number
+  sensor?: AntChannelId
   pressure?: number
   timestamp: string
 }
 export interface ParsedTankSummary {
-  sensor?: number
+  sensor?: AntChannelId
   start_pressure?: number
   end_pressure?: number
   volume_used?: number